refactor: structure using folders to prepare for test files
